Skiboot: Revert to legacy slot numbering on P9DSU
Some P9DSU systems (aka Boston) had incorrect attributes for some of the
slots, something that went without symptoms considering the disposition
of factory-configured adapters in those slots. But that was eventually
fixed anyway
(https://lists.ozlabs.org/pipermail/skiboot-stable/2019-March/000005.html)

One of the symptoms of this fix is that numbering of PCI slots may be
different before and after said update, depending on the disposition of
unpopulated slots in the systems.

This may cause "regressions" in the sense that systems pre-configured to
look for a particular adapter in a particular PCI bus may be confused
and fail.

The patch below addresses this problem by reverting to the previous
incorrect (but "working") behavior.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>