openpower-pnor: Use op940-specific pnor package for 24MB BOOTKERNEL
There have been a few requests to increase the BOOTKERNEL partition of
the PNOR to 24MB in order to fit more kernel drivers/firmware or more
rootfs binaries in Skiroot.

At the same time, there have been some discussions upstream on how to
address this in a generic way using Petitboot plugins.

As the requests so far have been only about IBM platforms using
release-op940 branch, and the upstream discussions are going in a
different way, I've opted to create a new release-op940 branch in the
already forked github.com/ibm-op-release/pnor repository to carry the
OP940-specific requirements for the PNOR layout.

This patch sets openpower-pnor.mk to that branch, making the
Mihawk-specific build-time patch with the same purpose obsolete.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>