Witherspoon: Enable Radeon on Skiroot/Petitboot
There have been requests to enable Skiroot/Petitboot console output
through Radeon cards (specifically Edwards2) on Witherspoon platforms.
So far this was not possible due to bugs in the Kernel that made
Edwards2 not kexec-safe.

Those bugs are now fixed, and we can now enable this for the Witherspoon
platform.

Note that not all Witherspoon systems will feature an Edwards2 adapter,
so in order to use it as a Skiroot/Petitboot console, the user needs to
manually change the Skiroot kernel command line and append commands that
instruct which console to use. On the Petitboot shell use EITHER:

 1) Command to blacklist the driver from the embedded AST graphics card:

    # nvram -p ibm,skiboot --update-config \
        bootargs="`cat /proc/cmdline` module_blacklist=ast"

 2) Command to map the Framebuffer console to the Radeon driver
    (considering AST graphics is still active):

    # nvram -p ibm,skiboot --update-config \
      bootargs="`cat /proc/cmdline` fbcon=map:N"

    Where "N" corresponds to the mapping function of currently active
    framebuffer devices (see 'Documentation/fb/fbcon.rst' in the Linux
    Kernel source for more information)

The user can revert the console output to the embedded AST graphics by
removing any 'module_blacklist=ast' or 'fbcon=map:1' entries from the
'bootargs' nvram setting, or by unsetting it to the default:

 # nvram -p ibm,skiboot --update-config bootargs=""

Since these changes are specific for Witherspoon, we're also not looking
to upstream those changes at this point in time.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>