p9dsu/kernel: Move to v5.10.50-openpower1
This moves the p9dsu platform to the v5.10 kernel. All of the secure
boot patches are now upstream, allowing them to be dropped.

This is manual backport of upsteam commit 86845e378c as it won't apply
cleanly on OP940 branch. This was done to support xfs boot partition.

Also backported upstream commit a823ae4f39c to p9dsu platform.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>