petitboot: Fix use-after-free error in DHCP handling
Backport the fix from open-power/petitboot/master that resolves
https://github.com/open-power/petitboot/issues/92. This can be reverted
when a petitboot v1.14 release is tagged and we update.

Alternatively, we could pull in this fix by advancing PETITBOOT_VERSION
to v1.13-14-g78f9f65fffd4, but moving from a tarball back to a git sha
is less straightforward as it also involves a revert of 6330437ee9b0
("petitboot: Update to v1.8.0 and switch to tarballs").

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>